+++ app/src/main/java/com/example/safenotepad/fragments/EditNoteFragment.kt
@@ -2,12 +2,10 @@ package com.example.safenotepad.fragments
 
 import android.content.Context
 import android.os.Bundle
-import android.text.method.PasswordTransformationMethod
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.EditText
 import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
 import androidx.navigation.fragment.findNavController
@@ -15,7 +13,7 @@ import com.example.safenotepad.databinding.FragmentEditNoteBinding
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.MutableLiveData
 import com.example.safenotepad.R
-import com.example.safenotepad.data.EncryptedSharedPreferencesDataStorage
+import com.example.safenotepad.data.sharedPreferences.EncryptedSharedPreferencesDataStorage
 import com.example.safenotepad.SharedViewModel
 import javax.crypto.Cipher
