+++ app/src/main/java/com/example/safenotepad/recyclerView/BindingAdapter.kt
@@ -2,7 +2,7 @@ package com.example.safenotepad.recyclerView
 
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.example.safenotepad.dao.Note
+import com.example.safenotepad.data.database.Note
 
 @BindingAdapter("listData")
 fun bindRecyclerView(recyclerView: RecyclerView, data: List<Note>?) {
